pyproject.toml
[project]
name = "harrisonai-dicomweb-cog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pydicom>=3.0.0",
    "dicomweb-client>=0.60.0",
    "google-auth>=2.0.0",
    "cog>=0.17.2",
    "torchvision>=0.26.0",
]
